Fixed guesses list removal logic
In the previous logic, it was assumed that the player already new the
answer to the game. In this way, the removed elements were the elements
that would not have the same correct and regular elements, given the
known answer.
However, the elements to be removed are any guesses that would not
produce that response if the last player's guess was the answer.